requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
google-cloud-aiplatform==1.68.0
google-cloud-firestore==2.16.0
pyjwt==2.9.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
email-validator==2.1.1
